requirements.txt
aiogram
aiogram_dialog
APScheduler
asyncpg
asyncssh
environs
nats-py
ormsgpack
greenlet
pydantic
SQLAlchemy
pytz
yookassa-async
fastapi
uvicorn

